requirements.txt
numpy==2.2.2
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.4.2
PyYAML==6.0.2
pytest==8.3.4
